+++ CHANGES.md
@@ -43,6 +43,7 @@ Changes in 11.4
 - Add "Student Info tab fields after" action hook: Add your own fields in Actions.php & Other_Info.inc.php
 - Add "User Info tab fields after" action hook: Add your own fields in Actions.php & Other_Info.inc.php
 - Fix NoInput() display '-' when $value is false in Inputs.php
+- Add "Report a bug" link to Resources in rosariosis.sql, rosariosis_mysql.sql, rosariosis_es.sql & rosariosis_fr.sql
 
 Changes in 11.3.3
 -----------------